# Conflicts:
#	src/main/resources/templates/add_vacation_point.html
#	src/main/resources/templates/admin/admin_panel.html

+++ src/main/java/dci/j24e01/TravelBlog/repositories/PendingLocationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PendingLocationRepository extends JpaRepository<PendingLocation, Integer> {
+public interface PendingLocationRepository extends JpaRepository<PendingLocation, Long> {
 
 }
